DIV-3594 - refactored submit AOS response call
Moved code to sendAosResponse to case orchestration service

Removed `if` check around env
