+++ services/caseOrchestration.js
@@ -49,7 +49,21 @@ const linkCase = req => {
     });
 };
 
+const sendAosResponse = (req, body) => {
+  const referenceNumber = req.session.referenceNumber;
+  const uri = `${CONF.services.caseOrchestration.submitAosUrl}/${referenceNumber}`;
+  const authTokenString = '__auth-token';
+  const headers = { Authorization: `${req.cookies[authTokenString]}` };
+
+  return request.post({ uri, body, headers, json: true })
+    .catch(error => {
+      logger.error(`Trying to connect to Case orchestration service error: ${error}`);
+      throw error;
+    });
+};
+
 module.exports = {
   getPetition,
-  linkCase
+  linkCase,
+  sendAosResponse
 };
